+++ src/app/app.module.ts
@@ -4,21 +4,21 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { FirstComponent } from './first/first.component';
-import { ProductComponent } from './product-list/components/product/product.component';
-import { ProductListComponent } from './product-list/components/product-list/product-list.component';
-import { CartListComponent } from './cart-list/components/cart-list/cart-list.component';
+import { CartModule } from './cart-list/cart.module';
+import { ProductsModule } from './product-list/products.module';
+import { SharedModule } from './shared/shared.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    FirstComponent,
-    ProductComponent,
-    ProductListComponent,
-    CartListComponent
+    FirstComponent
   ],
   imports: [
     BrowserModule,
-    AppRoutingModule
+    AppRoutingModule,
+    CartModule,
+    ProductsModule,
+    SharedModule
   ],
   providers: [],
   bootstrap: [AppComponent]