+++ src/app/cart-list/cart.module.ts
@@ -1,0 +1,23 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { CartListComponent } from './components/cart-list/cart-list.component';
+import { SharedModule } from '../shared/shared.module';
+import { CartItemComponent } from './components/cart-item/cart-item.component';
+import { FormsModule } from '@angular/forms';
+
+@NgModule({
+  declarations: [
+    CartListComponent,
+    CartItemComponent
+  ],
+  imports: [
+    CommonModule,
+    FormsModule,
+    SharedModule
+  ],
+  exports: [
+    CartListComponent
+  ]
+})
+export class CartModule { }