+++ src/app/cart-list/models/cart-item.model.ts
@@ -1,0 +1,8 @@
+import { ProductModel } from "src/app/shared/models/product.model";
+
+export class CartItemModel {
+    constructor(
+        public product: ProductModel,
+        public quantity: number
+    ) { }
+}