+++ src/app/product-list/components/product-list/product-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from "@angular/core";
 
 import { ProductModel } from "../../../shared/models/product.model";
 import { ProductService } from "../../services/product.service";
+import { CartService } from "src/app/cart-list/services/cart.service";
 
 @Component({
   selector: "app-product-list",
@@ -12,9 +13,16 @@ export class ProductListComponent implements OnInit {
 
   products: ProductModel[] = [];
 
-  constructor(private productService: ProductService) { }
+  constructor(
+    private productService: ProductService,
+    private cartService: CartService
+    ) { }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
+  }
+
+  onAddToCart(product: ProductModel): void {
+    this.cartService.addToCart(product);
   }
 }