+++ src/app/product-list/services/product.service.ts
@@ -20,11 +20,11 @@ export class ProductService {
 
   private createProducts(): ProductModel[] {
     return [
-      new ProductModel("Belt", "A belt for cats", 10, Category.Animals, true),
-      new ProductModel("T-shirt", "Women t-shirt", 30, Category.Clothes, true),
-      new ProductModel("Sandals", "Kids summer sandals", 35, Category.Clothes, false),
-      new ProductModel("Lamp", "A bedroom warm-light lamp", 50, Category.Home, true),
-      new ProductModel("Mascara", "L'oreal black mascara", 15, Category.Beauty, true)
+      new ProductModel(1, "Belt", "A belt for cats", 10, Category.Animals, true),
+      new ProductModel(2, "T-shirt", "Women t-shirt", 30, Category.Clothes, true),
+      new ProductModel(3, "Sandals", "Kids summer sandals", 35, Category.Clothes, false),
+      new ProductModel(4, "Lamp", "A bedroom warm-light lamp", 50, Category.Home, true),
+      new ProductModel(5, "Mascara", "L'oreal black mascara", 15, Category.Beauty, true)
     ]
   }
 }